# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Versus Api",
    "version": "1.0.0",
    "description": "Documentação do sistema de torneios Versus"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/team": {
      "get": {
        "summary": "Lista todos os times",
        "tags": [
          "Times"
        ],
        "responses": {
          "200": {
            "description": "Lista de times"
          }
        }
      },
      "post": {
        "summary": "Cria um novo time",
        "tags": [
          "Times"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Time Alpha",
                  "description": "Time para competição local"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Time criado com sucesso"
          },
          "400": {
            "description": "Erro ao criar time"
          }
        }
      }
    },
    "/team/{id}": {
      "get": {
        "summary": "Retorna um time pelo ID",
        "tags": [
          "Times"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do time",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time encontrado"
          },
          "404": {
            "description": "Time não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza um time pelo ID",
        "tags": [
          "Times"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do time",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Time Alpha Atualizado",
                  "description": "Descrição atualizada"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Time atualizado"
          },
          "400": {
            "description": "Erro na atualização"
          },
          "403": {
            "description": "Sem permissão para editar"
          }
        }
      },
      "delete": {
        "summary": "Deleta um time pelo ID",
        "tags": [
          "Times"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do time",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Time deletado com sucesso"
          },
          "400": {
            "description": "Erro ao deletar"
          },
          "403": {
            "description": "Sem permissão para deletar"
          }
        }
      }
    },
    "/team/{id}/inscribe": {
      "post": {
        "summary": "Inscreve o usuário logado no time",
        "tags": [
          "Times"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do time",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário inscrito no time"
          },
          "400": {
            "description": "Erro ao inscrever"
          }
        }
      }
    },
    "/team/{id}/unsubscribe": {
      "post": {
        "summary": "Remove inscrição do usuário no time",
        "tags": [
          "Times"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do time",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "userId": 5
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inscrição removida"
          },
          "400": {
            "description": "Erro ao remover inscrição"
          },
          "403": {
            "description": "Sem permissão para remover inscrição de outro usuário"
          }
        }
      }
    },
    "/team/{id}/inscriptions": {
      "get": {
        "summary": "Lista todos os usuários inscritos no time",
        "tags": [
          "Times"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do time",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários inscritos"
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Retorna todos os usuários",
        "tags": [
          "Usuários"
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários"
          }
        }
      },
      "post": {
        "summary": "Cria um novo usuário",
        "tags": [
          "Usuários"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "username": "joao",
                  "email": "joao@email.com",
                  "password": "123456"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado"
          },
          "400": {
            "description": "Erro na criação"
          }
        }
      },
      "put": {
        "summary": "Atualiza dados do usuário autenticado",
        "tags": [
          "Usuários"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "username": "joaoatualizado",
                  "email": "joao@novoemail.com"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado"
          },
          "400": {
            "description": "Erro na atualização"
          }
        }
      },
      "delete": {
        "summary": "Deleta o usuário autenticado",
        "tags": [
          "Usuários"
        ],
        "responses": {
          "204": {
            "description": "Usuário deletado"
          },
          "400": {
            "description": "Erro ao deletar"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "Retorna um usuário pelo ID",
        "tags": [
          "Usuários"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    }
  },
...
}